Set assert parameter by default
With an option to turn it off.

We don’t include an option for assertuser (yet?), because the user name
isn’t included in the response to the OAuth request, and so we’d need to
make a separate network request to get it. If someone asks for it, we
can add it later (defaulting to false like in m3api-botpassword).