Support PKCE on Node 12-14
Fall back to generating the code verifier without crypto, and using it
as the code challenge directly without hashing, if crypto is not
available (which should only be on older Node versions); add an option
allowing users to require crypto support if they don’t expect to be
running on older Node versions.

I’m not sure this can be meaningfully unit-tested (though that might
also be laziness speaking, to be honest); however, we can
integration-test it by running CI also on Node 14. (The integration
tests don’t work on Node 12 for unrelated reasons – `npm ci` already
fails, it can’t download ChromeDriver – but the package itself should
still work there.)