Automatically refresh access token
Using the new support for custom error handlers in m3api, react to
authorization errors by refreshing the access token and (if there is
enough time) retrying the request.

We’re purely reacting to errors about expired access tokens, but never
checking how long MediaWiki tells us the access tokens are valid for or
preemptively refreshing them. The main motivation for this is to keep
the code simple; if we track when the access token might expire, we’ll
have to deal with real time and time zones, and also serialize and
deserialize this information; I don’t think it’s worth the added
complexity. But the m3api-oauth2/autoRefresh option is used to avoid
getting into a loop of infinite retries if the API for some reason keeps
responding with mwoauth-invalid-authorization errors (which might not be
exclusive to the “expired access token” condition, after all).

The integration test for this takes a very long time to run (we sleep
for four hours, as that’s the usual time-to-live of the access token;
I’ll try to remember to re-check that every year or so), so only run it
when the RUN_SLOW_TESTS environment variable is set. Put that into a
separate CI job so that the normal job finishes quickly.

Bump the dev dependency to the main branch of m3api, which has the
needed errorHandlers support. Both this and the peer dependency will be
bumped again after the next m3api release (which will definitely be
before the next m3api-oauth2 release).

To check whether retryUntil has passed or not, we need the performance
global; import add-performance-global.js in the unit tests, and assume
that all other users (including the integration tests) are using
m3api/node.js, which pulls it in already.