Run slow tests against local MediaWiki
Instead of sleeping for four hours (and change) to let the OAuth grant
expire, run the slow test against a local wiki where we configure the
timeout to be much shorter. To implement this, I created a new GitHub
action to set up MediaWiki, which may also be useful to others. The
tests have to support running against different MediaWiki instances via
a new environment variable.

Bug: T374562