Make initOAuthSession() idempotent
Previously, there was an implicit assumption that, after the call to
initOAuthSession(), the user immediately opens the returned URL. But of
course that’s not guaranteed: they may just load the tool again, or even
in another tab. If the tool wants to still show a login link, it would
have to save the authorization URL itself, but this is brittle – any
call to initOAuthSession() would reset the code verifier and therefore
invalidate the previous authorization URL. It’s better to just make the
function idempotent and only generate a fresh code verifier if we don’t
have one yet – then, tools can call initOAuthSession() each time they
want to show a login link, and don’t have to care whether it’s the first
call or not.