Migrate to Wikimedia GitLab
The primary Git forge for m3api-oauth2 is now Wikimedia GitLab. The
documentation stays on GitHub Pages for now, so the GitHub repository
isn’t archived yet (and will be updated with each new release).

The GitLab CI imports and activates the regular jobs from the
ci-templates and passes through the right environment variables for
testing against the Beta cluster. Note that we have to use the
releng/node18-test-browser image instead of the default
releng/node18-test (which doesn’t include a browser), and also install
the matching Chromedriver version (v120). At least this should mean
we’ll have to update Chromedriver much less frequently than before –
only when we bump the releng/nodeXX-* image version, I expect.

Additionally, a test-slow-job implements the test-slow job from GitHub
Actions, importing a GitLab CI equivalent of setup-mediawiki [1] from
the ci-templates. Again we have to use a custom image,
releng/node20-test-browser-php81-composer, in order to have both
Chromium+Chromedriver and PHP+Composer available in the container;
there’s no node18 version of this image with PHP 8.1 (and MediaWiki is
no longer compatible with PHP 8.0), but compatibility with Node 18 is
still ensured by the other CI job. As the node20 image contains a more
recent version of Chromium than the node18 image, and we can only have
one in package.json, install a compatible version via command in the
test-slow-job.

[1]: https://github.com/lucaswerkmeister/setup-mediawiki/