Fix GitHub Actions CI
This includes the regular Chromedriver update (v139), but also updates
the $MEDIAWIKI_FULL_SCRIPT_PATH for the migration of the Beta cluster
domain from wmflabs.org to wmcloud; otherwise the waitUntil() condition
in the “log in” hook of the browser tests will never be true (the URL
will never be the wmflabs one we expect).

It’s perhaps worth noting that Chromedriver v139 only claims
compatibility with Node 20 and above, while we still run CI on Node 18.
As it still seems to work in practice, and the whole repo including its
CI will soon move to GitLab anyway (T392290), I’m leaving it this way
for now.