Check out an older OAuth commit in CI
Recent changes in the OAuth extension appear to have caused a
performance degradation, such that the test-slow-job in CI (which runs
against a local MediaWiki) no longer finishes successfully. (The
failures are mostly timeouts on the PHP side, but sometimes OOM errors
instead.) Avoid the errors for now by checking out an older commit.
(Compatibility with latest OAuth is still mostly checked via test-job,
which runs against the Beta cluster and yet is apparently not affected
by the issue.)

Bug: T409637