Use m3api-rest to make REST API requests
This means we no longer rely on m3api’s internal network interface and
can therefore use the ordinary ^ semver operator in our peer dependency.
The conversion is fairly straightforward and happily results in a net
reduction of JS lines here (not counting m3api-rest, of course).

One little oddity is that we have to explicitly specify that the access
token (Authorization header) should *not* be sent with the request that
gets an access token; this is most important in refreshOAuthSession(),
where the request will otherwise certainly fail, but we also do it in
completeOAuthSession() just in case there’s an access token in the
options for whatever reason.

The unit tests don’t change at all: they still mock internalPost(),
which is now called through m3api-rest rather than directly. This
arguably pushes them in the direction of integration tests, but I quite
like it because I think it gives me more confidence that everything
still works here and m3api-rest is doing the right thing. (It also means
the tests will have to be updated when m3api’s internal network
interface changes, of course.) I don’t think we really have another
option, anyway: we can’t really mock postForJson() itself without Jest
or other weird hacks to Node.js’s import machinery, which I don’t want
to use.

I don’t really understand the "dev" changes in package-lock.json, but
npm insists on them, so whatever.

The “internal improvements” in CHANGES.md are intended to cover not just
this commit but also the CI improvements in the previous couple of
commits.